fix(pre-deploy) Only delete _ss_env file if it exists

+++ mysite/scripts/pre_deploy.php
@@ -17,6 +17,8 @@ if (is_ss4()) {
         $cmds[] = "Generating env file";
         include __DIR__ . '/generate_env.php';
     }
-    $cmd = "rm $newPath/_ss_environment.php";
-    `$cmd`;
+    if (file_exists("$newPath/_ss_environment.php")) {
+        $cmd = "rm $newPath/_ss_environment.php";
+        `$cmd`;
+    }
 }
